Do not modify the mask to avoid misleading 'changed' reporting.
New keys in <DIR> are readable by the turnserver because the
default ACLs of <DIR> are set like:

  setfacl -m d:u:turnserver:r <DIR>

However, a call like 'setfacl -m u:turnserver:r <KEY>' will modify
the mask of <KEY> by default, leading to a misleading notification
during the playbook run.  This modification of the mask is neighter
needed nor usefull and can be avoided using '-n' in setfacl.